// src/modules/broker_referred_leads/broker_referred_leads.enum.ts
export enum BrokerReferredLeadStatus {
  LEAD_SUBMITTED = 'lead_submitted',
  LEAD_ACCEPTED = 'lead_accepted',
  LEAD_REJECTED = 'lead_rejected',
  TOUR_BOOKED = 'tour_booked',
  TOUR_COMPLETED = 'tour_completed',
  CLIENT_DEAL_LOST = 'client_deal_lost',
  CLIENT_DEAL_SIGNED = 'client_deal_signed',
  CLIENT_PAYMENT_RECEIVED = 'client_payment_received',
  BROKER_PAYMENT_INITIATED = 'broker_payment_initiated',
  BROKER_PAYMENT_COMPLETED = 'broker_payment_completed',
}